# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: command-center
    description: Defensive security workload telemetry
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /command-center/status:
    get:
      operationId: getCommandCenterStatus
      tags: [command-center]
      summary: Read live command center telemetry
      description: Returns the current in-memory coordinator snapshot for the defensive workload dashboard.
      responses:
        "200":
          description: Current telemetry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandCenterStatus"
  /command-center/task-orders:
    get:
      operationId: getTaskOrderCatalog
      tags: [command-center]
      summary: Read safe task-order preview catalog
      description: Returns archive-inspired defensive lab tasks mapped to bounded local executors. Payment, deposits, external targets, and credential recovery are not supported.
      responses:
        "200":
          description: Preview catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskOrderCatalog"
  /command-center/task-orders/quote:
    post:
      operationId: quoteTaskOrder
      tags: [command-center]
      summary: Calculate a preview-only synthetic task-order quote
      description: Estimates synthetic credits and completion time from total items, projected solved percentage, and measured hash rate. It never creates an order or dispatches work.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskOrderQuoteInput"
      responses:
        "200":
          description: Preview quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskOrderQuote"
        "400":
          description: Invalid task order quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command-center/task-orders/queue:
    post:
      operationId: queueTaskOrder
      tags: [command-center]
      summary: Queue a safe task for local miners
      description: Pins an available safe task to the loopback-only local miner pool. Current local workers receive a fresh job and future local workers receive the same selected task until another task is queued.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskOrderQueueInput"
      responses:
        "200":
          description: Local task queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskOrderQueueResult"
        "400":
          description: Task cannot be queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command-center/settings:
    get:
      operationId: getCommandCenterSettings
      tags: [command-center]
      summary: Read persisted command center settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandCenterSettings"
    put:
      operationId: saveCommandCenterSettings
      tags: [command-center]
      summary: Persist command center difficulty, miner count, and GPU backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandCenterSettingsInput"
      responses:
        "200":
          description: Saved settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandCenterSettings"
        "400":
          description: Invalid settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command-center/demo:
    post:
      operationId: runCommandCenterDemo
      tags: [command-center]
      summary: Run one bounded defensive workload demo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoInput"
      responses:
        "200":
          description: Completed workload demo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResult"
        "400":
          description: Unknown workload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Cross-origin public-key workload request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command-center/dns/migrate:
    post:
      operationId: migrateCommandCenterDns
      tags: [command-center]
      summary: Migrate the controlled lab DNS record and query it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DnsMigrationInput"
      responses:
        "200":
          description: DNS record migrated and queried over UDP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResult"
        "400":
          description: Invalid lab address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command-center/exploit-detection:
    post:
      operationId: runCommandCenterExploitDetection
      tags: [command-center]
      summary: Analyze a site snapshot without executing exploit content
      description: Performs static indicator matching against a user-supplied HTML, JavaScript, or header snapshot. The target URL is treated as a label; the server does not contact it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExploitDetectionInput"
      responses:
        "200":
          description: Static analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExploitDetectionResult"
        "400":
          description: Invalid snapshot or target URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    TaskOrderTask:
      type: object
      required: [id, title, description, category, status, statusReason, sourceWorkloadId, executor, defaultItems, creditsPerThousand, estimatedSolvePercent, authorizationBoundary]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [available, not_available]
        statusReason:
          type: string
        sourceWorkloadId:
          type: string
        executor:
          type: string
        defaultItems:
          type: integer
        creditsPerThousand:
          type: integer
        estimatedSolvePercent:
          type: number
        authorizationBoundary:
          type: string
          enum: [loopback-only]
    TaskOrderCatalog:
      type: object
      required: [mode, depositRequired, dispatchMode, notice, defaultHashRate, tasks]
      properties:
        mode:
          type: string
          enum: [preview]
        depositRequired:
          type: boolean
          enum: [false]
        dispatchMode:
          type: string
          enum: [local-preview-only]
        notice:
          type: string
        defaultHashRate:
          type: integer
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskOrderTask"
    TaskOrderQuoteInput:
      type: object
      required: [taskId, hashRate, solvedPercent, totalItems]
      properties:
        taskId:
          type: string
          minLength: 1
        hashRate:
          type: integer
          minimum: 1
          maximum: 1000000000000000
          description: Hashes or bounded work units per second; measured telemetry is the default.
        solvedPercent:
          type: number
          minimum: 0
          maximum: 100
        totalItems:
          type: integer
          minimum: 1
          maximum: 100000000
    TaskOrderQuote:
      type: object
      required: [taskId, taskName, status, availabilityReason, hashRate, solvedPercent, totalItems, estimatedSolvedItems, estimatedSeconds, estimatedTime, balanceRequiredCredits, depositRequired, dispatchMode]
      properties:
        taskId:
          type: string
        taskName:
          type: string
        status:
          type: string
          enum: [available, not_available]
        availabilityReason:
          type: string
        hashRate:
          type: integer
        solvedPercent:
          type: number
        totalItems:
          type: integer
        estimatedSolvedItems:
          type: integer
        estimatedSeconds:
          type: integer
        estimatedTime:
          type: string
        balanceRequiredCredits:
          type: integer
        depositRequired:
          type: boolean
          enum: [false]
        dispatchMode:
          type: string
          enum: [local-preview-only]
    TaskOrderQueueInput:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: string
          minLength: 1
    TaskOrderQueueResult:
      type: object
      required: [taskId, taskName, sourceWorkloadId, status, activeMiners, jobIds, jobsIssued, dispatchMode, message]
      properties:
        taskId:
          type: string
        taskName:
          type: string
        sourceWorkloadId:
          type: string
        status:
          type: string
          enum: [queued]
        activeMiners:
          type: integer
        jobIds:
          type: array
          items:
            type: string
        jobsIssued:
          type: integer
        dispatchMode:
          type: string
          enum: [loopback-local]
        message:
          type: string
    CommandCenterStatus:
      type: object
      required:
        - generatedAt
        - coordinator
        - dns
        - workloads
        - ledger
      properties:
        generatedAt:
          type: string
          format: date-time
        coordinator:
          $ref: "#/components/schemas/CoordinatorStatus"
        dns:
          $ref: "#/components/schemas/DnsTelemetry"
        workloads:
          type: array
          items:
            $ref: "#/components/schemas/WorkloadTelemetry"
        ledger:
          type: array
          items:
            $ref: "#/components/schemas/LedgerBlock"
    CoordinatorStatus:
      type: object
      required: [status, activeMiners, configuredMiners, workerHealth, throughputHealth, gpuMining, intentionalScaleDowns, unexpectedRestarts, poolPort, webPort, policy, acceptedShares, jobsIssued, totalHashes, hashRate, difficulty, minerSource]
      properties:
        status:
          type: string
        activeMiners:
          type: integer
        configuredMiners:
          type: integer
        workerHealth:
          $ref: "#/components/schemas/LocalWorkerHealth"
        throughputHealth:
          $ref: "#/components/schemas/ThroughputHealth"
        gpuMining:
          $ref: "#/components/schemas/GpuMiningStatus"
        intentionalScaleDowns:
          type: integer
          description: Number of local worker threads intentionally retired while reducing the configured pool size
        unexpectedRestarts:
          type: integer
          description: Number of local worker thread exits that required an automatic replacement
        poolPort:
          type: integer
        webPort:
          type: integer
        policy:
          type: string
        acceptedShares:
          type: integer
        jobsIssued:
          type: integer
        totalHashes:
          type: integer
        hashRate:
          type: integer
          description: Measured hashes per second across all live miner workers
        difficulty:
          type: string
        minerSource:
          type: string
    LocalWorkerHealth:
      type: object
      required: [enabled, configured, running, healthy]
      properties:
        enabled:
          type: boolean
          description: Whether the local worker-thread pool is enabled
        configured:
          type: integer
          description: Number of local worker threads requested
        running:
          type: integer
          description: Number of local worker threads currently running and not being retired
        healthy:
          type: boolean
          description: Whether the local worker pool matches its configured size
    ThroughputHealth:
      type: object
      required: [status, releaseFloorHashRate, observedHashRate, acceptedShares, evidenceAgeMs, evidenceFresh, explanation]
      properties:
        status:
          type: string
          enum: [healthy, warming-up, below-floor, offline]
        releaseFloorHashRate:
          type: integer
          description: Minimum hash rate used as a bounded release-health indicator, not an exact speed target
        observedHashRate:
          type: integer
          description: Measured aggregate hash rate from local workers
        acceptedShares:
          type: integer
          description: Accepted proof count backing the live throughput signal
        evidenceAgeMs:
          type: integer
          nullable: true
          description: Age of the most recent accepted proof; null before the first proof
        evidenceFresh:
          type: boolean
        explanation:
          type: string
    GpuMiningStatus:
      type: object
      required: [enabled, requestedBackend, available, active, backend, device, processCount, stdoutLines, measuredHashRate, effectiveHashRateMhs, noncesChecked, lastEventAt, lastShareAt, explanation]
      properties:
        enabled:
          type: boolean
          description: Whether the configured GPU process backend is enabled
        requestedBackend:
          type: string
          enum: [auto, gpu, disabled]
        available:
          type: boolean
          description: Whether a supported local GPU was detected
        active:
          type: boolean
        backend:
          type: string
          enum: [gpu, cpu-fallback, disabled]
        device:
          type: string
        processCount:
          type: integer
        stdoutLines:
          type: integer
          description: Number of newline-delimited telemetry events read from the miner process
        measuredHashRate:
          type: integer
          description: Latest measured raw hashes per second reported by the miner process
        effectiveHashRateMhs:
          type: number
          description: Relay-scale effective rate in MH/s; telemetry only and not physical hardware throughput
        noncesChecked:
          type: integer
          description: Number of nonce candidates checked for the latest reported share
        lastEventAt:
          type: string
          format: date-time
          nullable: true
        lastShareAt:
          type: string
          format: date-time
          nullable: true
        explanation:
          type: string
    DnsTelemetry:
      type: object
      required: [status, domain, ip, ttl, port, queries, serial]
      properties:
        status:
          type: string
        domain:
          type: string
        ip:
          type: string
        ttl:
          type: integer
        port:
          type: integer
        queries:
          type: integer
        serial:
          type: integer
    WorkloadTelemetry:
      type: object
      required: [id, name, purpose, severity, category, algorithm, executionMode, executor, authorizationBoundary, evidencePolicy, status, completionState, progress, verifiedShares, lastProof, rate, hashes, hashRate, note, evidenceSource, lastEvidenceAt, evidenceDigest]
      properties:
        id:
          type: string
        name:
          type: string
        purpose:
          type: string
          description: What this bounded workload does
        severity:
          type: string
        category:
          type: string
          description: Safety-bounded workload category
        algorithm:
          type: string
          description: Concrete local algorithm used by the workload executor
        executionMode:
          type: string
          enum: [local-live]
          description: Execution uses bounded services, fixtures, and computation on loopback
        executor:
          type: string
          description: Stable local executor identity
        authorizationBoundary:
          type: string
          enum: [loopback-only]
        evidencePolicy:
          type: string
          enum: [executor-result-required, executor-not-implemented]
        status:
          type: string
        completionState:
          type: string
          enum: [waiting, running, evidence_pending, complete, rejected]
        progress:
          type: integer
        verifiedShares:
          type: integer
        lastProof:
          type: string
        rate:
          type: string
        hashes:
          type: integer
          description: Validated nonce hashes attempted for this workload
        hashRate:
          type: integer
          description: Measured hashes per second for this workload
        note:
          type: string
        evidenceSource:
          type: string
          enum: [none, local-executor, mining-telemetry, simulated, mock, fixture]
        lastEvidenceAt:
          type: string
          format: date-time
          nullable: true
        evidenceDigest:
          type: string
        acceptedReport:
          $ref: "#/components/schemas/AcceptedWorkloadReport"
    LedgerBlock:
      type: object
      required: [height, task, severity, hash, nonce, miner, timestamp]
      properties:
        height:
          type: integer
        task:
          type: string
        severity:
          type: string
        hash:
          type: string
        nonce:
          type: integer
        miner:
          type: string
        timestamp:
          type: string
          format: date-time
    DemoInput:
      type: object
      required: [workloadId]
      properties:
        workloadId:
          type: string
          minLength: 1
        mnemonicWords:
          type: integer
          enum: [12, 24]
          default: 24
          description: Word count for the w-01 mnemonic fixture
        difficulty:
          type: string
          description: Leading-zero hexadecimal difficulty for the local miner job
          enum: ["00", "000", "0000"]
        publicKeyArtifact:
          type: string
          minLength: 1
          maxLength: 16384
          description: Supplied PEM, base64-DER, or hex-DER public-key artifact for the public-key validation workload. Key bytes are parsed in memory and never retained.
    CommandCenterSettingsInput:
      type: object
      required: [difficulty, minerCount, gpuBackend]
      properties:
        difficulty:
          type: string
          enum: ["00", "000", "0000"]
        minerCount:
          type: integer
          minimum: 1
          maximum: 16
        gpuBackend:
          type: string
          enum: [auto, gpu, disabled]
    CommandCenterSettings:
      type: object
      required: [difficulty, minerCount, gpuBackend]
      properties:
        difficulty:
          type: string
          enum: ["00", "000", "0000"]
        minerCount:
          type: integer
          minimum: 1
          maximum: 16
        gpuBackend:
          type: string
          enum: [auto, gpu, disabled]
    DnsMigrationInput:
      type: object
      required: [targetIp]
      properties:
        targetIp:
          type: string
          description: One of the two reserved lab addresses
          enum: ["127.0.0.1", "127.0.0.2"]
        difficulty:
          type: string
          description: Leading-zero hexadecimal difficulty for the local miner validation job
          enum: ["00", "000", "0000"]
    ExploitDetectionInput:
      type: object
      required: [targetUrl, snapshot, snapshotKind]
      properties:
        targetUrl:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
          description: Authorized site label. It is never contacted by this operation.
        snapshot:
          type: string
          minLength: 1
          maxLength: 200000
          description: Pasted HTML, JavaScript, or response headers to analyze statically.
        snapshotKind:
          type: string
          enum: [html, javascript, headers, combined]
          description: The kind of artifact supplied in snapshot.
    ExploitDetectionFinding:
      type: object
      required: [id, title, severity, confidence, matchedSignals, description, recommendation]
      properties:
        id:
          type: string
        title:
          type: string
        severity:
          type: string
          enum: [Critical, High, Medium, Low]
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        matchedSignals:
          type: array
          items:
            type: string
        description:
          type: string
        recommendation:
          type: string
    ExploitDetectionResult:
      type: object
      required: [runId, status, targetUrl, scannedAt, artifactBytes, artifactSha256, analysisMode, findings, evidence, safety]
      properties:
        runId:
          type: string
        status:
          type: string
          enum: [detected, not_detected, insufficient_evidence]
        targetUrl:
          type: string
          format: uri
        scannedAt:
          type: string
          format: date-time
        artifactBytes:
          type: integer
        artifactSha256:
          type: string
        analysisMode:
          type: string
        findings:
          type: array
          items:
            $ref: "#/components/schemas/ExploitDetectionFinding"
        evidence:
          type: array
          items:
            type: string
        safety:
          type: string
    DemoDetail:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
    AcceptedWorkloadReport:
      type: object
      description: Latest accepted result from the real bounded executor; raw input artifacts are never returned. This is shared coordinator state, not proof of request ownership.
      required: [answer, details, evidence, executionEvidence]
      properties:
        answer:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/DemoDetail"
        evidence:
          type: array
          items:
            type: string
        executionEvidence:
          $ref: "#/components/schemas/WorkloadEvidence"
    DemoResult:
      type: object
      required: [runId, status, workloadId, workloadName, severity, startedAt, finishedAt, answer, details, evidence, executionEvidence, evidenceAccepted, evidenceRejectionReason, safety, mining]
      properties:
        runId:
          type: string
        status:
          type: string
        workloadId:
          type: string
        workloadName:
          type: string
        severity:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        answer:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/DemoDetail"
        evidence:
          type: array
          items:
            type: string
        executionEvidence:
          $ref: "#/components/schemas/WorkloadEvidence"
        evidenceAccepted:
          type: boolean
        evidenceRejectionReason:
          type: string
        safety:
          type: string
        mining:
          $ref: "#/components/schemas/MiningTelemetry"
        dnsSerial:
          type: integer
          description: Authoritative DNS record serial after a controlled migration
        dnsTarget:
          type: string
          description: Resolved loopback target after a controlled migration
        dnsRedirect:
          $ref: "#/components/schemas/DnsRedirectResult"
        dnsRecordProof:
          $ref: "#/components/schemas/DnsRecordProof"
    DnsRedirectResult:
      type: object
      required: [statusCode, url, body]
      properties:
        statusCode:
          type: integer
        url:
          type: string
          format: uri
        body:
          type: string
    DnsRecordProof:
      type: object
      required: [canonical, recordHash, keyFingerprint, signature]
      properties:
        canonical:
          type: string
        recordHash:
          type: string
        keyFingerprint:
          type: string
        signature:
          type: string
    MiningTelemetry:
      type: object
      required: [status, workloadId, jobId, executor, evidenceRole, inputDigest, activeMiners, acceptedShares, difficulty, jobsIssued, hashes, hashRate, effectiveHashRateMhs, workloadShares, lastHash, lastMiner]
      properties:
        status:
          type: string
        workloadId:
          type: string
        jobId:
          type: string
        executor:
          type: string
        evidenceRole:
          type: string
          enum: [mining-telemetry-only]
          description: Mining telemetry is attribution evidence and cannot complete a workload
        inputDigest:
          type: string
          description: Digest of the issued workload job identity, not proof of executor completion
        activeMiners:
          type: integer
        acceptedShares:
          type: integer
        difficulty:
          type: string
        jobsIssued:
          type: integer
        hashes:
          type: integer
          description: Validated nonce hashes attempted for the selected workload
        hashRate:
          type: integer
          description: Measured hashes per second for the selected workload
        effectiveHashRateMhs:
          type: number
          description: Relay-scale effective rate in MH/s for demo telemetry; not physical hardware throughput
        workloadShares:
          type: integer
        lastHash:
          type: string
        lastMiner:
          type: string
    WorkloadEvidence:
      type: object
      required: [workloadId, jobId, executor, boundedPartition, source, realWork, inputDigest, artifactDigest, evidenceDigest, observedAt, completionState]
      properties:
        workloadId:
          type: string
        jobId:
          type: string
        executor:
          type: string
        boundedPartition:
          type: string
          description: Bounded partition identity used by the named executor
        source:
          type: string
          enum: [local-executor, mining-telemetry, simulated, mock, fixture]
        realWork:
          type: boolean
        inputDigest:
          type: string
        artifactDigest:
          type: string
          description: SHA-256 digest of the supplied artifact or executor input
        evidenceDigest:
          type: string
        observedAt:
          type: string
          format: date-time
        completionState:
          type: string
          enum: [complete]
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string