# lib/api-zod/src/generated/types/commandCenterSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommandCenterSettingsDifficulty } from './commandCenterSettingsDifficulty';
import type { CommandCenterSettingsGpuBackend } from './commandCenterSettingsGpuBackend';

export interface CommandCenterSettings {
  difficulty: CommandCenterSettingsDifficulty;
  /**
     * @minimum 1
     * @maximum 16
     */
  minerCount: number;
  gpuBackend: CommandCenterSettingsGpuBackend;
}