# lib/api-zod/src/generated/types/ledgerBlock.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LedgerBlock {
  height: number;
  task: string;
  severity: string;
  hash: string;
  nonce: number;
  miner: string;
  timestamp: Date;
}